Refactor map speed handling and update state machine wait time
- Remove unused `wmap_getMapScaledSpeed` function and its references
- Simplify `wmap_getMapScaledTime` function to directly use `globalSpeed`
- Update projectile speed calculations to use raw speed values
- Change `waitEndGameTime` from `f32` to `f64` for better precision
- Adjust wait state creation to directly calculate `waitEndGameTime`

+++ src/war_state_machine.h
@@ -51,7 +51,7 @@ struct _WarState
 
         struct
         {
-            f32 waitEndGameTime;
+            f64 waitEndGameTime;
         } wait;
 
         struct
